Separate reward claim functions as hooks for each ecosystem
