+++ apps/extension/src/hooks/claim/index.ts
@@ -1,0 +1,3 @@
+export * from "./use-claim-all-each-state";
+export * from "./use-cosmos-claim-rewards";
+export * from "./use-starknet-claim-rewards";
