+++ apps/extension/src/pages/main/components/claim-all/index.tsx
@@ -1,4 +1,4 @@
-import React, { FunctionComponent, useEffect, useRef, useState } from "react";
+import React, { FunctionComponent, useEffect, useState } from "react";
 import { Column, Columns } from "../../../../components/column";
 import { Button } from "../../../../components/button";
 import { Stack } from "../../../../components/stack";
@@ -15,28 +15,9 @@ import {
 } from "../../../../components/icon";
 import { observer } from "mobx-react-lite";
 import { useStore } from "../../../../stores";
-import { CoinPretty, Dec, Int, PricePretty } from "@keplr-wallet/unit";
-import {
-  AminoSignResponse,
-  AppCurrency,
-  BroadcastMode,
-  Coin,
-  FeeCurrency,
-  ModularChainInfo,
-  StdSignDoc,
-} from "@keplr-wallet/types";
-import { InExtensionMessageRequester } from "@keplr-wallet/router-extension";
-import { BACKGROUND_PORT } from "@keplr-wallet/router";
-import {
-  PrivilegeCosmosSignAminoWithdrawRewardsMsg,
-  PrivilegeStarknetSignClaimRewardsMsg,
-  SendTxMsg,
-} from "@keplr-wallet/background";
-import { action, makeObservable, observable } from "mobx";
+import { CoinPretty, Dec, PricePretty } from "@keplr-wallet/unit";
+import { ModularChainInfo } from "@keplr-wallet/types";
 import { Tooltip } from "../../../../components/tooltip";
-import { isSimpleFetchError } from "@keplr-wallet/simple-fetch";
-import { useNotification } from "../../../../hooks/notification";
-import { useNavigate } from "react-router";
 import { Skeleton } from "../../../../components/skeleton";
 import { YAxis } from "../../../../components/axis";
 import Color from "color";
@@ -44,16 +25,24 @@ import { SpecialButton } from "../../../../components/special-button";
 import { Gutter } from "../../../../components/gutter";
 import { FormattedMessage, useIntl } from "react-intl";
 import { CurrencyImageFallback } from "../../../../components/image";
-import { DefaultGasPriceStep } from "@keplr-wallet/hooks";
-import { IChainInfoImpl, MakeTxResponse } from "@keplr-wallet/stores";
-import { Call, CallData, num } from "starknet";
+import {
+  ClaimAllEachState,
+  useCosmosClaimRewards,
+  useClaimAllEachState,
+  useStarknetClaimRewards,
+} from "../../../../hooks/claim";
 
 interface ViewClaimToken extends Omit<ViewToken, "chainInfo"> {
   modularChainInfo: ModularChainInfo;
   price?: PricePretty;
   onClaimAll: (
     chainId: string,
-    rewardToken: CoinPretty
+    rewardToken: CoinPretty,
+    state: ClaimAllEachState
+  ) => void | Promise<void>;
+  onClaimSingle: (
+    chainId: string,
+    state: ClaimAllEachState
   ) => void | Promise<void>;
 }
 
@@ -110,466 +99,6 @@ const Styles = {
   `,
 };
 
-// XXX: 좀 이상하긴 한데 상위/하위 컴포넌트가 state를 공유하기 쉽게하려고 이렇게 한다...
-class ClaimAllEachState {
-  @observable
-  isLoading: boolean = false;
-
-  @observable
-  failedReason: Error | undefined = undefined;
-
-  constructor() {
-    makeObservable(this);
-  }
-
-  @action
-  setIsLoading(value: boolean): void {
-    this.isLoading = value;
-  }
-
-  @action
-  setFailedReason(value: Error | undefined): void {
-    this.isLoading = false;
-    this.failedReason = value;
-  }
-}
-
-const useClaimAllEachState = () => {
-  const { chainStore } = useStore();
-  const statesRef = useRef(new Map<string, ClaimAllEachState>());
-  const getClaimAllEachState = (chainId: string): ClaimAllEachState => {
-    // modular chain의 경우 chainIdentifier가 없다.
-    // 따라서 chainId를 사용한다.
-
-    const chainIdentifier = chainStore.hasChain(chainId)
-      ? chainStore.getChain(chainId).chainIdentifier
-      : chainId;
-
-    let state = statesRef.current.get(chainIdentifier);
-    if (!state) {
-      state = new ClaimAllEachState();
-      statesRef.current.set(chainIdentifier, state);
-    }
-
-    return state;
-  };
-
-  return {
-    states: statesRef.current.values(),
-    getClaimAllEachState,
-  };
-};
-
-const usePrepareCosmosInnerTx = () => {
-  const { queriesStore, priceStore } = useStore();
-  const intl = useIntl();
-
-  const prepareFeeCurrency = async (
-    chainInfo: IChainInfoImpl,
-    bech32Address: string
-  ) => {
-    let feeCurrency = chainInfo.hasFeature("feemarket")
-      ? undefined
-      : chainInfo.feeCurrencies.find(
-          (cur) =>
-            cur.coinMinimalDenom === chainInfo.stakeCurrency?.coinMinimalDenom
-        );
-
-    if (chainInfo.hasFeature("osmosis-base-fee-beta") && feeCurrency) {
-      const queryBaseFee = queriesStore.get(chainInfo.chainId).osmosis
-        .queryBaseFee;
-      const queryRemoteBaseFeeStep = queriesStore.simpleQuery.queryGet<{
-        low?: number;
-        average?: number;
-        high?: number;
-      }>(
-        "https://gjsttg7mkgtqhjpt3mv5aeuszi0zblbb.lambda-url.us-west-2.on.aws/osmosis/osmosis-base-fee-beta.json"
-      );
-
-      await queryBaseFee.waitFreshResponse();
-      await queryRemoteBaseFeeStep.waitFreshResponse();
-
-      const baseFee = queryBaseFee.baseFee;
-      const remoteBaseFeeStep = queryRemoteBaseFeeStep.response;
-      if (baseFee) {
-        const low = remoteBaseFeeStep?.data.low
-          ? parseFloat(
-              baseFee.mul(new Dec(remoteBaseFeeStep.data.low)).toString(8)
-            )
-          : feeCurrency.gasPriceStep?.low ?? DefaultGasPriceStep.low;
-        const average = Math.max(
-          low,
-          remoteBaseFeeStep?.data.average
-            ? parseFloat(
-                baseFee.mul(new Dec(remoteBaseFeeStep.data.average)).toString(8)
-              )
-            : feeCurrency.gasPriceStep?.average ?? DefaultGasPriceStep.average
-        );
-        const high = Math.max(
-          average,
-          remoteBaseFeeStep?.data.high
-            ? parseFloat(
-                baseFee.mul(new Dec(remoteBaseFeeStep.data.high)).toString(8)
-              )
-            : feeCurrency.gasPriceStep?.high ?? DefaultGasPriceStep.high
-        );
-
-        feeCurrency = {
-          ...feeCurrency,
-          gasPriceStep: {
-            low,
-            average,
-            high,
-          },
-        };
-      }
-    }
-
-    if (!feeCurrency) {
-      let prev:
-        | {
-            balance: CoinPretty;
-            price: PricePretty | undefined;
-          }
-        | undefined;
-
-      const feeCurrencies = await (async () => {
-        if (chainInfo.hasFeature("feemarket")) {
-          const queryFeeMarketGasPrices = queriesStore.get(chainInfo.chainId)
-            .cosmos.queryFeeMarketGasPrices;
-          await queryFeeMarketGasPrices.waitFreshResponse();
-
-          const result: FeeCurrency[] = [];
-
-          for (const gasPrice of queryFeeMarketGasPrices.gasPrices) {
-            const currency = await chainInfo.findCurrencyAsync(gasPrice.denom);
-            if (currency) {
-              let multiplication = {
-                low: 1.1,
-                average: 1.2,
-                high: 1.3,
-              };
-
-              const multificationConfig = queriesStore.simpleQuery.queryGet<{
-                [str: string]:
-                  | {
-                      low: number;
-                      average: number;
-                      high: number;
-                    }
-                  | undefined;
-              }>(
-                "https://gjsttg7mkgtqhjpt3mv5aeuszi0zblbb.lambda-url.us-west-2.on.aws",
-                "/feemarket/info.json"
-              );
-
-              if (multificationConfig.response) {
-                const _default =
-                  multificationConfig.response.data["__default__"];
-                if (
-                  _default &&
-                  _default.low != null &&
-                  typeof _default.low === "number" &&
-                  _default.average != null &&
-                  typeof _default.average === "number" &&
-                  _default.high != null &&
-                  typeof _default.high === "number"
-                ) {
-                  multiplication = {
-                    low: _default.low,
-                    average: _default.average,
-                    high: _default.high,
-                  };
-                }
-                const specific =
-                  multificationConfig.response.data[chainInfo.chainIdentifier];
-                if (
-                  specific &&
-                  specific.low != null &&
-                  typeof specific.low === "number" &&
-                  specific.average != null &&
-                  typeof specific.average === "number" &&
-                  specific.high != null &&
-                  typeof specific.high === "number"
-                ) {
-                  multiplication = {
-                    low: specific.low,
-                    average: specific.average,
-                    high: specific.high,
-                  };
-                }
-              }
-
-              result.push({
-                ...currency,
-                gasPriceStep: {
-                  low: parseFloat(
-                    new Dec(multiplication.low).mul(gasPrice.amount).toString()
-                  ),
-                  average: parseFloat(
-                    new Dec(multiplication.average)
-                      .mul(gasPrice.amount)
-                      .toString()
-                  ),
-                  high: parseFloat(
-                    new Dec(multiplication.high).mul(gasPrice.amount).toString()
-                  ),
-                },
-              });
-            }
-          }
-
-          return result;
-        } else {
-          return chainInfo.feeCurrencies;
-        }
-      })();
-      for (const chainFeeCurrency of feeCurrencies) {
-        const currency = await chainInfo.findCurrencyAsync(
-          chainFeeCurrency.coinMinimalDenom
-        );
-        if (currency) {
-          const queries = queriesStore.get(chainInfo.chainId);
-
-          const balance = queries.queryBalances
-            .getQueryBech32Address(bech32Address)
-            .getBalance(currency);
-          if (balance && balance.balance.toDec().gt(new Dec(0))) {
-            const price = await priceStore.waitCalculatePrice(
-              balance.balance,
-              "usd"
-            );
-
-            if (!prev) {
-              feeCurrency = {
-                ...chainFeeCurrency,
-                ...currency,
-              };
-              prev = {
-                balance: balance.balance,
-                price,
-              };
-            } else {
-              if (!prev.price) {
-                if (prev.balance.toDec().lt(balance.balance.toDec())) {
-                  feeCurrency = {
-                    ...chainFeeCurrency,
-                    ...currency,
-                  };
-                  prev = {
-                    balance: balance.balance,
-                    price,
-                  };
-                }
-              } else if (price) {
-                if (prev.price.toDec().lt(price.toDec())) {
-                  feeCurrency = {
-                    ...chainFeeCurrency,
-                    ...currency,
-                  };
-                  prev = {
-                    balance: balance.balance,
-                    price,
-                  };
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-
-    return feeCurrency;
-  };
-
-  const estimateFee = async (
-    chainInfo: IChainInfoImpl,
-    feeCurrency: FeeCurrency,
-    tx: MakeTxResponse
-  ) => {
-    const simulated = await tx.simulate();
-
-    // Gas adjustment is 1.5
-    // Since there is currently no convenient way to adjust the gas adjustment on the UI,
-    // Use high gas adjustment to prevent failure.
-    const gasEstimated = new Dec(simulated.gasUsed * 1.5).truncate();
-    let fee = {
-      denom: feeCurrency.coinMinimalDenom,
-      amount: new Dec(feeCurrency.gasPriceStep?.average ?? 0.025)
-        .mul(new Dec(gasEstimated))
-        .roundUp()
-        .toString(),
-    };
-
-    // USD 기준으로 average fee가 0.2달러를 넘으면 low로 설정해서 보낸다.
-    const averageFeePrice = await priceStore.waitCalculatePrice(
-      new CoinPretty(feeCurrency, fee.amount),
-      "usd"
-    );
-    if (averageFeePrice && averageFeePrice.toDec().gte(new Dec(0.2))) {
-      fee = {
-        denom: feeCurrency.coinMinimalDenom,
-        amount: new Dec(feeCurrency.gasPriceStep?.low ?? 0.025)
-          .mul(new Dec(gasEstimated))
-          .roundUp()
-          .toString(),
-      };
-      console.log(
-        `(${chainInfo.chainId}) Choose low gas price because average fee price is greater or equal than 0.2 USD`
-      );
-    }
-
-    // Ensure fee currency fetched before querying balance
-    const feeCurrencyFetched = await chainInfo.findCurrencyAsync(
-      feeCurrency.coinMinimalDenom
-    );
-    if (!feeCurrencyFetched) {
-      throw new Error(
-        intl.formatMessage({
-          id: "error.can-not-find-balance-for-fee-currency",
-        })
-      );
-    }
-
-    return { gasEstimated, fee, feeCurrencyFetched };
-  };
-
-  const checkBalance = async (
-    chainInfo: IChainInfoImpl,
-    targetToken: CoinPretty,
-    fee: Coin,
-    feeCurrencyFetched: AppCurrency,
-    bech32Address: string
-  ) => {
-    const queries = queriesStore.get(chainInfo.chainId);
-    const balance = queries.queryBalances
-      .getQueryBech32Address(bech32Address)
-      .getBalance(feeCurrencyFetched);
-
-    if (!balance) {
-      throw new Error(
-        intl.formatMessage({
-          id: "error.can-not-find-balance-for-fee-currency",
-        })
-      );
-    }
-
-    await balance.waitResponse();
-
-    if (new Dec(balance.balance.toCoin().amount).lt(new Dec(fee.amount))) {
-      throw new Error(
-        intl.formatMessage({
-          id: "error.not-enough-balance-to-pay-fee",
-        })
-      );
-    }
-
-    if (
-      (targetToken.toCoin().denom === fee.denom &&
-        new Dec(targetToken.toCoin().amount).lte(new Dec(fee.amount))) ||
-      (await (async () => {
-        if (targetToken.toCoin().denom !== fee.denom) {
-          if (
-            targetToken.currency.coinGeckoId &&
-            feeCurrencyFetched.coinGeckoId
-          ) {
-            const rewardPrice = await priceStore.waitCalculatePrice(
-              targetToken,
-              "usd"
-            );
-            const feePrice = await priceStore.waitCalculatePrice(
-              new CoinPretty(feeCurrencyFetched, fee.amount),
-              "usd"
-            );
-            if (
-              rewardPrice &&
-              rewardPrice.toDec().gt(new Dec(0)) &&
-              feePrice &&
-              feePrice.toDec().gt(new Dec(0))
-            ) {
-              if (rewardPrice.toDec().mul(new Dec(1.2)).lte(feePrice.toDec())) {
-                return true;
-              }
-            }
-          }
-        }
-
-        return false;
-      })())
-    ) {
-      console.log(
-        `(${chainInfo.chainId}) Skip claim rewards. Fee: ${fee.amount}${
-          fee.denom
-        } is greater than stakable reward: ${targetToken.toCoin().amount}${
-          targetToken.toCoin().denom
-        }`
-      );
-      throw new Error(
-        intl.formatMessage({
-          id: "error.claimable-reward-is-smaller-than-the-required-fee",
-        })
-      );
-    }
-  };
-
-  const execute = async (
-    tx: MakeTxResponse,
-    gasEstimated: Int,
-    fee: Coin,
-    onBroadcasted?: (txHash: Uint8Array) => void,
-    onFulfill?: (tx: any) => void
-  ) => {
-    await tx.send(
-      {
-        gas: gasEstimated.toString(),
-        amount: [fee],
-      },
-      "",
-      {
-        signAmino: async (
-          chainId: string,
-          signer: string,
-          signDoc: StdSignDoc
-        ): Promise<AminoSignResponse> => {
-          const requester = new InExtensionMessageRequester();
-
-          return await requester.sendMessage(
-            BACKGROUND_PORT,
-            new PrivilegeCosmosSignAminoWithdrawRewardsMsg(
-              chainId,
-              signer,
-              signDoc
-            )
-          );
-        },
-        sendTx: async (
-          chainId: string,
-          tx: Uint8Array,
-          mode: BroadcastMode
-        ): Promise<Uint8Array> => {
-          const requester = new InExtensionMessageRequester();
-
-          return await requester.sendMessage(
-            BACKGROUND_PORT,
-            new SendTxMsg(chainId, tx, mode, true)
-          );
-        },
-      },
-      {
-        onBroadcasted,
-        onFulfill,
-      }
-    );
-  };
-
-  return {
-    prepareFeeCurrency,
-    estimateFee,
-    checkBalance,
-    execute,
-  };
-};
-
 const zeroDec = new Dec(0);
 
 export const ClaimAll: FunctionComponent<{ isNotReady?: boolean }> = observer(
@@ -579,7 +108,6 @@ export const ClaimAll: FunctionComponent<{ isNotReady?: boolean }> = observer(
       chainStore,
       accountStore,
       queriesStore,
-      starknetAccountStore,
       starknetQueriesStore,
       priceStore,
       keyRingStore,
@@ -591,427 +119,10 @@ export const ClaimAll: FunctionComponent<{ isNotReady?: boolean }> = observer(
     const [isExpanded, setIsExpanded] = useState(false);
     const { states, getClaimAllEachState } = useClaimAllEachState();
 
-    const { prepareFeeCurrency, estimateFee, checkBalance, execute } =
-      usePrepareCosmosInnerTx();
-
-    const handleCosmosInnerClaim = (
-      chainId: string,
-      rewardToken: CoinPretty
-    ) => {
-      const cosmosChainInfo = chainStore.getChain(chainId);
-      const account = accountStore.getAccount(chainId);
-      if (!account.bech32Address) {
-        return;
-      }
-
-      const queries = queriesStore.get(chainId);
-      const queryRewards = queries.cosmos.queryRewards.getQueryBech32Address(
-        account.bech32Address
-      );
-
-      const validatorAddresses =
-        queryRewards.getDescendingPendingRewardValidatorAddresses(
-          account.isNanoLedger ? 5 : 8
-        );
-
-      if (validatorAddresses.length === 0) {
-        return;
-      }
-
-      const state = getClaimAllEachState(chainId);
-
-      state.setIsLoading(true);
-
-      const tx =
-        account.cosmos.makeWithdrawDelegationRewardTx(validatorAddresses);
-
-      (async () => {
-        // feemarket feature가 있는 경우 이후의 로직에서 사용할 수 있는 fee currency를 찾아야하기 때문에 undefined로 시작시킨다.
-        const feeCurrency = await prepareFeeCurrency(
-          cosmosChainInfo,
-          account.bech32Address
-        );
-
-        if (feeCurrency) {
-          try {
-            const { gasEstimated, fee, feeCurrencyFetched } = await estimateFee(
-              cosmosChainInfo,
-              feeCurrency,
-              tx
-            );
-
-            await checkBalance(
-              cosmosChainInfo,
-              rewardToken,
-              fee,
-              feeCurrencyFetched,
-              account.bech32Address
-            );
-
-            const onBroadcasted = () => {
-              analyticsStore.logEvent("complete_claim_all", {
-                chainId: cosmosChainInfo.chainId,
-                chainName: cosmosChainInfo.chainName,
-              });
-            };
-
-            const onFulfill = (tx: any) => {
-              // Tx가 성공한 이후에 rewards가 다시 쿼리되면서 여기서 빠지는게 의도인데...
-              // 쿼리하는 동안 시간차가 있기 때문에 훼이크로 그냥 1초 더 기다린다.
-              setTimeout(() => {
-                state.setIsLoading(false);
-              }, 1000);
-
-              if (tx.code) {
-                state.setFailedReason(new Error(tx["raw_log"]));
-              }
-            };
-
-            await execute(tx, gasEstimated, fee, onBroadcasted, onFulfill);
-          } catch (e) {
-            if (isSimpleFetchError(e) && e.response) {
-              const response = e.response;
-              if (
-                response.status === 400 &&
-                response.data?.message &&
-                typeof response.data.message === "string" &&
-                response.data.message.includes("invalid empty tx")
-              ) {
-                state.setFailedReason(
-                  new Error(
-                    intl.formatMessage({
-                      id: "error.outdated-cosmos-sdk",
-                    })
-                  )
-                );
-                return;
-              }
-            }
-
-            state.setFailedReason(e);
-            console.log(e);
-            return;
-          }
-        } else {
-          state.setFailedReason(
-            new Error(
-              intl.formatMessage({
-                id: "error.can-not-find-fee-for-claim-all",
-              })
-            )
-          );
-          return;
-        }
-      })();
-    };
-
-    const handleStarknetInnerClaim = (
-      chainId: string,
-      rewardToken: CoinPretty
-    ) => {
-      const starknetChainInfo = chainStore.getModularChain(chainId);
-      const account = accountStore.getAccount(chainId);
-      if (!account.starknetHexAddress) {
-        return;
-      }
-
-      const starknetAccount = starknetAccountStore.getAccount(chainId);
-
-      const starknetQueries = starknetQueriesStore.get(chainId);
-      const queryValidators = starknetQueries.queryValidators;
-      const validators = queryValidators.validators;
-      const queryStakingInfo = queryValidators
-        .getQueryPoolMemberInfoMap(account.starknetHexAddress)
-        ?.getQueryStakingInfo(validators);
-
-      const claimableRewards =
-        queryStakingInfo?.getDescendingPendingClaimableRewards(
-          account.isNanoLedger ? 5 : 8
-        );
-
-      if (!claimableRewards || claimableRewards.length === 0) {
-        return;
-      }
-
-      const state = getClaimAllEachState(chainId);
-
-      state.setIsLoading(true);
-
-      // build tx
-      const calls: Call[] = [];
-
-      for (const claimableReward of claimableRewards) {
-        if (claimableReward.poolAddress) {
-          calls.push({
-            contractAddress: claimableReward.poolAddress,
-            calldata: CallData.compile([account.starknetHexAddress]),
-            entrypoint: "claim_rewards",
-          });
-        }
-      }
-
-      const currencies = chainStore
-        .getModularChainInfoImpl(chainId)
-        .getCurrencies("starknet");
-
-      const STRK = currencies.find((c) => c.coinDenom === "STRK");
-      const ETH = currencies.find((c) => c.coinDenom === "ETH");
-
-      (async () => {
-        try {
-          // select fee currency
-          let feeCurrency = STRK || ETH;
-          if (STRK && ETH) {
-            // compare the price of balances of STRK and ETH
-            const querySTRK =
-              starknetQueries.queryStarknetERC20Balance.getBalance(
-                chainId,
-                chainStore,
-                account.starknetHexAddress,
-                STRK?.coinMinimalDenom
-              );
-            const queryETH =
-              starknetQueries.queryStarknetERC20Balance.getBalance(
-                chainId,
-                chainStore,
-                account.starknetHexAddress,
-                ETH?.coinMinimalDenom
-              );
-
-            const strkBalance = querySTRK?.balance;
-            const ethBalance = queryETH?.balance;
-
-            if (strkBalance && ethBalance) {
-              const strkPrice = await priceStore.waitCalculatePrice(
-                strkBalance,
-                "usd"
-              );
-              const ethPrice = await priceStore.waitCalculatePrice(
-                ethBalance,
-                "usd"
-              );
-
-              if (strkPrice && ethPrice) {
-                if (strkPrice.sub(ethPrice).toDec().lt(zeroDec)) {
-                  feeCurrency = STRK;
-                }
-              }
-            }
-          }
-
-          if (!feeCurrency) {
-            throw new Error(
-              intl.formatMessage({
-                id: "error.can-not-find-fee-for-claim-all",
-              })
-            );
-          }
-
-          // estimate fee
-          const {
-            gas_consumed,
-            data_gas_consumed,
-            gas_price,
-            overall_fee,
-            resourceBounds,
-            unit,
-          } = await starknetAccount.estimateInvokeFee(
-            account.starknetHexAddress,
-            calls,
-            feeCurrency.coinDenom === "STRK" ? "STRK" : "ETH"
-          );
-
-          const gasMargin = new Dec(1.2);
-          const gasPriceMargin = new Dec(1.5);
-
-          const gasConsumedDec = new Dec(gas_consumed);
-          const dataGasConsumedDec = new Dec(data_gas_consumed);
-          const sigVerificationGasConsumedDec = new Dec(583);
-          const totalGasConsumed = gasConsumedDec
-            .add(dataGasConsumedDec)
-            .add(sigVerificationGasConsumedDec);
-
-          const gasPriceDec = new Dec(gas_price);
-          // overallFee는 (gas_consumed * gas_price + data_gas_consumed * data_gas_price)로 계산됨.
-          const overallFeeDec = new Dec(overall_fee);
-
-          const sigVerificationFee =
-            sigVerificationGasConsumedDec.mul(gasPriceDec);
-
-          const adjustedOverallFee = overallFeeDec.add(sigVerificationFee);
-
-          const adjustedGasPrice = adjustedOverallFee.quo(totalGasConsumed);
-
-          const gasPrice = new CoinPretty(feeCurrency, adjustedGasPrice);
-
-          let fee:
-            | {
-                version: "v1" | "v3";
-                currency: AppCurrency;
-                gasPrice: CoinPretty;
-                maxGasPrice: CoinPretty;
-                gas: Dec;
-                maxGas: Dec;
-              }
-            | undefined;
-
-          if (feeCurrency.coinDenom === "ETH" && unit === "WEI") {
-            const maxGasPriceForETH = gasPrice.mul(gasPriceMargin);
-            const maxGasForETH = totalGasConsumed.mul(gasMargin);
-
-            fee = {
-              version: "v1",
-              currency: feeCurrency,
-              gasPrice,
-              maxGasPrice: maxGasPriceForETH,
-              gas: totalGasConsumed,
-              maxGas: maxGasForETH,
-            };
-          } else if (feeCurrency.coinDenom === "STRK" && unit === "FRI") {
-            const l1Gas = resourceBounds.l1_gas;
-
-            const maxGasForSTRK = adjustedOverallFee
-              .quo(gasPriceDec)
-              .mul(gasMargin);
-            const maxGasPriceForSTRK = gasPrice.mul(gasPriceMargin);
-
-            const maxPricePerUnitForSTRK = new CoinPretty(
-              feeCurrency,
-              num.hexToDecimalString(l1Gas.max_price_per_unit)
-            );
-
-            const finalMaxGasPriceForSTRK = maxPricePerUnitForSTRK
-              .sub(maxGasPriceForSTRK)
-              .toDec()
-              .gt(new Dec(0))
-              ? maxPricePerUnitForSTRK
-              : maxGasPriceForSTRK;
-
-            fee = {
-              version: "v3",
-              currency: feeCurrency,
-              gasPrice,
-              maxGasPrice: finalMaxGasPriceForSTRK,
-              gas: totalGasConsumed,
-              maxGas: maxGasForSTRK,
-            };
-          }
-
-          if (!fee) {
-            throw new Error(
-              intl.formatMessage({
-                id: "error.can-not-find-fee-for-claim-all",
-              })
-            );
-          }
-
-          // compare the account balance and fee
-          const feeCurrencyBalance =
-            starknetQueries.queryStarknetERC20Balance.getBalance(
-              chainId,
-              chainStore,
-              account.starknetHexAddress,
-              feeCurrency.coinMinimalDenom
-            );
-
-          if (
-            !feeCurrencyBalance ||
-            feeCurrencyBalance.balance
-              .toDec()
-              .lt(fee.maxGasPrice.mul(fee.gas).toDec())
-          ) {
-            throw new Error(
-              intl.formatMessage({
-                id: "error.claimable-reward-is-smaller-than-the-required-fee",
-              })
-            );
-          }
-
-          // compare the price of claimable rewards and fee (just consider maxGasPrice)
-          const maxFee = fee.maxGasPrice.mul(fee.gas);
-
-          const rewardsPrice = await priceStore.waitCalculatePrice(
-            rewardToken,
-            "usd"
-          );
-          const maxFeePrice = await priceStore.waitCalculatePrice(
-            maxFee,
-            "usd"
-          );
-
-          if (!rewardsPrice || !maxFeePrice) {
-            throw new Error(
-              intl.formatMessage({
-                id: "error.can-not-find-price-for-claim-all",
-              })
-            );
-          }
-
-          // if the price of claimable rewards is less than the price of fee, throw an error
-          if (rewardsPrice.sub(maxFeePrice).toDec().lt(zeroDec)) {
-            throw new Error(
-              intl.formatMessage({
-                id: "error.claimable-reward-is-smaller-than-the-required-fee",
-              })
-            );
-          }
-
-          // execute
-          const { transaction_hash: txHash } = await starknetAccount.execute(
-            account.starknetHexAddress,
-            calls,
-            fee.version === "v3"
-              ? {
-                  type: "STRK",
-                  gas: fee.gas.roundUp().toString(),
-                  maxGasPrice: fee.maxGasPrice
-                    .mul(new Dec(10 ** fee.currency.coinDecimals))
-                    .toDec()
-                    .roundUp()
-                    .toString(),
-                }
-              : {
-                  type: "ETH",
-                  maxFee: fee.maxGasPrice
-                    .mul(fee.maxGas)
-                    .mul(new Dec(10 ** fee.currency.coinDecimals))
-                    .toDec()
-                    .roundUp()
-                    .toString(),
-                },
-            async (chainId, calls, details) => {
-              const requester = new InExtensionMessageRequester();
-
-              return await requester.sendMessage(
-                BACKGROUND_PORT,
-                new PrivilegeStarknetSignClaimRewardsMsg(
-                  chainId,
-                  calls,
-                  details
-                )
-              );
-            }
-          );
-
-          if (!txHash) {
-            throw new Error("Failed to claim all");
-          }
-
-          analyticsStore.logEvent("complete_claim_all", {
-            chainId: starknetChainInfo.chainId,
-            chainName: starknetChainInfo.chainName,
-          });
-
-          setTimeout(() => {
-            state.setIsLoading(false);
-          }, 1000);
-        } catch (e) {
-          state.setFailedReason(e);
-          console.log(e);
-          return;
-        }
-      })();
-    };
+    const { handleCosmosClaimAllEach, handleCosmosClaimSingle } =
+      useCosmosClaimRewards();
+    const { handleStarknetClaimAllEach, handleStarknetClaimSingle } =
+      useStarknetClaimRewards();
 
     const viewClaimTokens: ViewClaimToken[] = (() => {
       const res: ViewClaimToken[] = [];
@@ -1050,7 +161,8 @@ export const ClaimAll: FunctionComponent<{ isNotReady?: boolean }> = observer(
                   modularChainInfo: modularChainInfo,
                   isFetching: queryRewards.isFetching,
                   error: queryRewards.error,
-                  onClaimAll: handleCosmosInnerClaim,
+                  onClaimAll: handleCosmosClaimAllEach,
+                  onClaimSingle: handleCosmosClaimSingle,
                 });
               }
             }
@@ -1082,7 +194,8 @@ export const ClaimAll: FunctionComponent<{ isNotReady?: boolean }> = observer(
               modularChainInfo: starknetChainInfo,
               isFetching: queryStakingInfo?.isFetching ?? false,
               error: queryValidators?.error, // ignore queryStakingInfo error
-              onClaimAll: handleStarknetInnerClaim,
+              onClaimAll: handleStarknetClaimAllEach,
+              onClaimSingle: handleStarknetClaimSingle,
             });
           }
         }
@@ -1163,9 +276,14 @@ export const ClaimAll: FunctionComponent<{ isNotReady?: boolean }> = observer(
       }
 
       for (const viewClaimToken of viewClaimTokens) {
+        const state = getClaimAllEachState(
+          viewClaimToken.modularChainInfo.chainId
+        );
+
         viewClaimToken.onClaimAll(
           viewClaimToken.modularChainInfo.chainId,
-          viewClaimToken.token
+          viewClaimToken.token,
+          state
         );
       }
     };
@@ -1371,127 +489,13 @@ const ViewCosmosClaimTokenItem: FunctionComponent<{
   state: ClaimAllEachState;
   itemsLength: number;
 }> = observer(({ viewClaimToken, state, itemsLength }) => {
-  const { analyticsStore, accountStore, queriesStore } = useStore();
-
-  const intl = useIntl();
-  const navigate = useNavigate();
-  const notification = useNotification();
-
-  const [isSimulating, setIsSimulating] = useState(false);
-
-  // TODO: Add below property to config.ui.ts
-  const defaultGasPerDelegation = 140000;
-
-  const claim = async () => {
-    analyticsStore.logEvent("click_claim", {
-      chainId: viewClaimToken.modularChainInfo.chainId,
-      chainName: viewClaimToken.modularChainInfo.chainName,
-    });
-
-    if (state.failedReason) {
-      state.setFailedReason(undefined);
-      return;
-    }
-    const chainId = viewClaimToken.modularChainInfo.chainId;
-    const account = accountStore.getAccount(chainId);
-
-    const queries = queriesStore.get(chainId);
-    const queryRewards = queries.cosmos.queryRewards.getQueryBech32Address(
-      account.bech32Address
-    );
-
-    const validatorAddresses =
-      queryRewards.getDescendingPendingRewardValidatorAddresses(
-        account.isNanoLedger ? 5 : 8
-      );
-
-    if (validatorAddresses.length === 0) {
-      return;
-    }
-
-    const tx =
-      account.cosmos.makeWithdrawDelegationRewardTx(validatorAddresses);
-
-    let gas = new Int(validatorAddresses.length * defaultGasPerDelegation);
-
-    try {
-      setIsSimulating(true);
-
-      const simulated = await tx.simulate();
-
-      // Gas adjustment is 1.5
-      // Since there is currently no convenient way to adjust the gas adjustment on the UI,
-      // Use high gas adjustment to prevent failure.
-      gas = new Dec(simulated.gasUsed * 1.5).truncate();
-    } catch (e) {
-      console.log(e);
-    }
-
-    // TODO: gas price step이 고정되어있지 않은 경우에 대해서 처리 해야함 ex) osmosis-base-fee-beta
-
-    try {
-      await tx.send(
-        {
-          gas: gas.toString(),
-          amount: [],
-        },
-        "",
-        {},
-        {
-          onBroadcasted: () => {
-            analyticsStore.logEvent("complete_claim", {
-              chainId: viewClaimToken.modularChainInfo.chainId,
-              chainName: viewClaimToken.modularChainInfo.chainName,
-            });
-          },
-          onFulfill: (tx: any) => {
-            if (tx.code != null && tx.code !== 0) {
-              console.log(tx.log ?? tx.raw_log);
-              notification.show(
-                "failed",
-                intl.formatMessage({ id: "error.transaction-failed" }),
-                ""
-              );
-              return;
-            }
-            notification.show(
-              "success",
-              intl.formatMessage({
-                id: "notification.transaction-success",
-              }),
-              ""
-            );
-          },
-        }
-      );
-
-      navigate("/", {
-        replace: true,
-      });
-    } catch (e) {
-      if (e?.message === "Request rejected") {
-        return;
-      }
-
-      console.log(e);
-      notification.show(
-        "failed",
-        intl.formatMessage({ id: "error.transaction-failed" }),
-        ""
-      );
-      navigate("/", {
-        replace: true,
-      });
-    } finally {
-      setIsSimulating(false);
-    }
-  };
+  const { accountStore } = useStore();
 
   const isLoading =
     accountStore.getAccount(viewClaimToken.modularChainInfo.chainId)
       .isSendingMsg === "withdrawRewards" ||
     state.isLoading ||
-    isSimulating;
+    state.isSimulating;
 
   return (
     <ViewClaimTokenItemContent
@@ -1499,7 +503,12 @@ const ViewCosmosClaimTokenItem: FunctionComponent<{
       state={state}
       itemsLength={itemsLength}
       isLoading={isLoading}
-      onClick={claim}
+      onClick={() => {
+        viewClaimToken.onClaimSingle(
+          viewClaimToken.modularChainInfo.chainId,
+          state
+        );
+      }}
     />
   );
 });
@@ -1509,164 +518,13 @@ const ViewStarknetClaimTokenItem: FunctionComponent<{
   state: ClaimAllEachState;
   itemsLength: number;
 }> = observer(({ viewClaimToken, state, itemsLength }) => {
-  const {
-    accountStore,
-    starknetAccountStore,
-    starknetQueriesStore,
-    analyticsStore,
-    chainStore,
-  } = useStore();
-
-  const intl = useIntl();
-  const navigate = useNavigate();
-  const notification = useNotification();
-  const [isSimulating, setIsSimulating] = useState(false);
-
-  const claim = async () => {
-    analyticsStore.logEvent("click_claim", {
-      chainId: viewClaimToken.modularChainInfo.chainId,
-      chainName: viewClaimToken.modularChainInfo.chainName,
-    });
-
-    if (state.failedReason) {
-      state.setFailedReason(undefined);
-      return;
-    }
-
-    const chainId = viewClaimToken.modularChainInfo.chainId;
-    const account = accountStore.getAccount(chainId);
-    const starknetAccount = starknetAccountStore.getAccount(chainId);
-
-    const queries = starknetQueriesStore.get(chainId);
-    const queryValidators = queries.queryValidators;
-
-    const validators = queryValidators.validators;
-    const queryStakingInfo = queryValidators
-      .getQueryPoolMemberInfoMap(account.starknetHexAddress)
-      ?.getQueryStakingInfo(validators);
-
-    const claimableRewards =
-      queryStakingInfo?.getDescendingPendingClaimableRewards(
-        account.isNanoLedger ? 5 : 8
-      );
-
-    if (!claimableRewards || claimableRewards.length === 0) {
-      return;
-    }
-
-    const currencies = chainStore
-      .getModularChainInfoImpl(chainId)
-      .getCurrencies("starknet");
-
-    const feeCurrency = currencies.find(
-      (currency) => currency.coinDenom === "ETH"
-    );
-
-    if (!feeCurrency) {
-      return;
-    }
-
-    const calls: Call[] = [];
-
-    for (const claimableReward of claimableRewards) {
-      if (claimableReward.poolAddress) {
-        calls.push({
-          contractAddress: claimableReward.poolAddress,
-          calldata: CallData.compile([account.starknetHexAddress]),
-          entrypoint: "claim_rewards",
-        });
-      }
-    }
-
-    try {
-      setIsSimulating(true);
-
-      // estimate fee with ETH first
-      const { gas_consumed, data_gas_consumed, gas_price, overall_fee, unit } =
-        await starknetAccount.estimateInvokeFee(
-          account.starknetHexAddress,
-          calls,
-          "ETH"
-        );
-
-      if (unit !== "WEI") {
-        throw new Error("Invalid fee unit");
-      }
-
-      const gasMargin = new Dec(1.2);
-      const gasPriceMargin = new Dec(1.5);
-
-      const gasConsumedDec = new Dec(gas_consumed);
-      const dataGasConsumedDec = new Dec(data_gas_consumed);
-      const sigVerificationGasConsumedDec = new Dec(583);
-      const totalGasConsumed = gasConsumedDec
-        .add(dataGasConsumedDec)
-        .add(sigVerificationGasConsumedDec);
-
-      const gasPriceDec = new Dec(gas_price);
-      const overallFeeDec = new Dec(overall_fee);
-
-      const sigVerificationFee = sigVerificationGasConsumedDec.mul(gasPriceDec);
-
-      const adjustedOverallFee = overallFeeDec.add(sigVerificationFee);
-
-      const adjustedGasPrice = adjustedOverallFee.quo(totalGasConsumed);
-
-      const gasPrice = new CoinPretty(feeCurrency, adjustedGasPrice);
-
-      const maxGasPrice = gasPrice.mul(gasPriceMargin);
-      const maxGas = totalGasConsumed.mul(gasMargin);
-
-      const { transaction_hash: txHash } = await starknetAccount.execute(
-        account.starknetHexAddress,
-        calls,
-        {
-          type: "ETH",
-          maxFee: maxGasPrice
-            .mul(maxGas)
-            .mul(new Dec(10 ** feeCurrency.coinDecimals))
-            .toDec()
-            .roundUp()
-            .toString(),
-        }
-      );
-
-      if (!txHash) {
-        throw new Error("Failed to claim rewards");
-      }
-
-      analyticsStore.logEvent("complete_claim", {
-        chainId: viewClaimToken.modularChainInfo.chainId,
-        chainName: viewClaimToken.modularChainInfo.chainName,
-      });
-
-      navigate("/", {
-        replace: true,
-      });
-    } catch (e) {
-      if (e?.message === "Request rejected") {
-        return;
-      }
-
-      console.log(e);
-      notification.show(
-        "failed",
-        intl.formatMessage({ id: "error.transaction-failed" }),
-        ""
-      );
-      navigate("/", {
-        replace: true,
-      });
-    } finally {
-      setIsSimulating(false);
-    }
-  };
+  const { starknetAccountStore } = useStore();
 
   const isLoading =
-    state.isLoading ||
     starknetAccountStore.getAccount(viewClaimToken.modularChainInfo.chainId)
       .isSendingTx ||
-    isSimulating;
+    state.isSimulating ||
+    state.isLoading;
 
   return (
     <ViewClaimTokenItemContent
@@ -1674,7 +532,12 @@ const ViewStarknetClaimTokenItem: FunctionComponent<{
       state={state}
       itemsLength={itemsLength}
       isLoading={isLoading}
-      onClick={claim}
+      onClick={() => {
+        viewClaimToken.onClaimSingle(
+          viewClaimToken.modularChainInfo.chainId,
+          state
+        );
+      }}
     />
   );
 });
